# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerCopilot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  userId: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  currentTitle?: string | null;
  /** @nullable */
  yearsOfExperience?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileInput {
  fullName?: string;
  email?: string;
  phone?: string;
  location?: string;
  linkedinUrl?: string;
  websiteUrl?: string;
  bio?: string;
  currentTitle?: string;
  yearsOfExperience?: number;
}

export interface Resume {
  id: number;
  userId: string;
  title: string;
  content: string;
  isDefault?: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ResumeInput {
  title: string;
  content: string;
  isDefault?: boolean;
}

export interface ResumeUpdate {
  title?: string;
  content?: string;
  isDefault?: boolean;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  saved: "saved",
  applied: "applied",
  interview: "interview",
  offer: "offer",
  rejected: "rejected",
  withdrawn: "withdrawn",
} as const;

export interface Application {
  id: number;
  userId: string;
  companyName: string;
  jobTitle: string;
  /** @nullable */
  jobUrl?: string | null;
  /** @nullable */
  jobDescription?: string | null;
  status: ApplicationStatus;
  /** @nullable */
  appliedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  resumeId?: number | null;
  /** @nullable */
  matchScore?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type ApplicationInputStatus =
  (typeof ApplicationInputStatus)[keyof typeof ApplicationInputStatus];

export const ApplicationInputStatus = {
  saved: "saved",
  applied: "applied",
  interview: "interview",
  offer: "offer",
  rejected: "rejected",
  withdrawn: "withdrawn",
} as const;

export interface ApplicationInput {
  companyName: string;
  jobTitle: string;
  jobUrl?: string;
  jobDescription?: string;
  status?: ApplicationInputStatus;
  appliedAt?: string;
  notes?: string;
  resumeId?: number;
}

export type ApplicationUpdateStatus =
  (typeof ApplicationUpdateStatus)[keyof typeof ApplicationUpdateStatus];

export const ApplicationUpdateStatus = {
  saved: "saved",
  applied: "applied",
  interview: "interview",
  offer: "offer",
  rejected: "rejected",
  withdrawn: "withdrawn",
} as const;

export interface ApplicationUpdate {
  companyName?: string;
  jobTitle?: string;
  jobUrl?: string;
  jobDescription?: string;
  status?: ApplicationUpdateStatus;
  appliedAt?: string;
  notes?: string;
  resumeId?: number;
  matchScore?: number;
}

export type GeneratedResponseType =
  (typeof GeneratedResponseType)[keyof typeof GeneratedResponseType];

export const GeneratedResponseType = {
  resume_summary: "resume_summary",
  bullet_points: "bullet_points",
  cover_letter: "cover_letter",
  why_role: "why_role",
  tell_me_about_yourself: "tell_me_about_yourself",
  why_hire_you: "why_hire_you",
  describe_experience: "describe_experience",
  strengths: "strengths",
} as const;

export interface GeneratedResponse {
  id: number;
  applicationId: number;
  type: GeneratedResponseType;
  content: string;
  createdAt: string;
}

export type AiGenerateInputTypesItem =
  (typeof AiGenerateInputTypesItem)[keyof typeof AiGenerateInputTypesItem];

export const AiGenerateInputTypesItem = {
  resume_summary: "resume_summary",
  bullet_points: "bullet_points",
  cover_letter: "cover_letter",
  why_role: "why_role",
  tell_me_about_yourself: "tell_me_about_yourself",
  why_hire_you: "why_hire_you",
  describe_experience: "describe_experience",
  strengths: "strengths",
} as const;

export interface AiGenerateInput {
  applicationId: number;
  resumeContent: string;
  jobDescription: string;
  types: AiGenerateInputTypesItem[];
}

export interface AiGenerateResult {
  results: GeneratedResponse[];
}

export interface AtsAnalysisInput {
  resumeContent: string;
  jobDescription: string;
}

export interface AtsAnalysisResult {
  matchScore: number;
  matchedKeywords: string[];
  missingKeywords: string[];
  suggestions: string[];
}

export interface QuestionAnswerInput {
  question: string;
  resumeContent: string;
  jobDescription: string;
  applicationId?: number;
}

export interface QuestionAnswerResult {
  answer: string;
}

export type DashboardStatsByStatus = {
  saved?: number;
  applied?: number;
  interview?: number;
  offer?: number;
  rejected?: number;
  withdrawn?: number;
};

export interface DashboardStats {
  totalApplications: number;
  byStatus: DashboardStatsByStatus;
  totalResumes: number;
  totalGeneratedResponses: number;
  recentActivityCount: number;
}

export type ListApplicationsParams = {
  status?: string;
};