# lib/api-zod/src/generated/types/aiGenerateResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerCopilot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedResponse } from "./generatedResponse";

export interface AiGenerateResult {
  results: GeneratedResponse[];
}