# lib/api-zod/src/generated/types/generatedResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerCopilot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedResponseType } from "./generatedResponseType";

export interface GeneratedResponse {
  id: number;
  applicationId: number;
  type: GeneratedResponseType;
  content: string;
  createdAt: string;
}