# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerCopilot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  userId: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  currentTitle?: string | null;
  /** @nullable */
  yearsOfExperience?: number | null;
  createdAt: string;
  updatedAt: string;
}