# lib/api-zod/src/generated/types/questionAnswerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerCopilot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionAnswerInput {
  question: string;
  resumeContent: string;
  jobDescription: string;
  applicationId?: number;
}